Renamed the shim command to exec
