+++ cmd/state/internal/cmdtree/exec.go
@@ -6,24 +6,24 @@ import (
 	"github.com/ActiveState/cli/internal/captain"
 	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/primer"
-	"github.com/ActiveState/cli/internal/runners/shim"
+	"github.com/ActiveState/cli/internal/runners/exec"
 )
 
-func newShimCommand(prime *primer.Values, args ...string) *captain.Command {
-	runner := shim.New(prime)
+func newExecCommand(prime *primer.Values, args ...string) *captain.Command {
+	runner := exec.New(prime)
 
-	params := shim.NewParams()
+	params := exec.NewParams()
 
 	cmd := captain.NewCommand(
-		"shim",
+		"exec",
 		"",
-		locale.T("shim_description"),
+		locale.T("exec_description"),
 		prime.Output(),
 		prime.Config(),
 		[]*captain.Flag{
 			{
 				Name:        "path",
-				Description: locale.Tl("flag_state_shim_path_description", "Path to project that is providing the default environment."),
+				Description: locale.Tl("flag_state_exec_path_description", "Path to project that is providing the default environment."),
 				Value:       &params.Path,
 			},
 		},
@@ -32,14 +32,13 @@ func newShimCommand(prime *primer.Values, args ...string) *captain.Command {
 			return runner.Run(params, args...)
 		},
 	)
-	cmd.SetHidden(true)
 	cmd.SetSkipChecks(true)
 	cmd.SetDeferAnalytics(true)
 
 	// Cobra will handle the `--` delimiter if flag parsing is enabled.
 	// If the delimeter is not present we have to disable flag parsing
-	// to ensure flags are passed to the shimmed command rather than
-	// parsed as a flag for `state shim`
+	// to ensure flags are passed to the command rather than
+	// parsed as a flag for `state exec`
 	if !strings.Contains(strings.Join(args, " "), " -- ") {
 		cmd.SetDisableFlagParsing(true)
 	}
